First working server hand made (finally...)
I mean there certainly are lot of copied code snippets in it
from various sources, but that is definitely not a copy paste
of a tutorial code.
It's much more of a frankenstein code made from multiple tutorials
and random tokio based project code bases, knitted together as I
tried to guess my way into a working server knowing the basics of
tokio (having read the official tokio tutorial earlier this year).

Oh and "RTFM"ing helped a lot to solve the type errors, but it felt
like a huge documentation labyrinth (tokio + future docs).