Source structure refactoring (flatter)
I don't remember why I started to add a level of hierarchy but it just
complicated the reading and compartimentalize the informations which is
part of what I ranted about the Rust documentation. So let's not do that
unless it is necessary :)

Also it will make the README chain more maintanable (only 1 README
linking to its own sections).

Oh and it will also truthfully suggest that each example is independant
and therefore that any client can interact with any server as long as
they use the same codec.